Move OracleContract contract interface to the contracts package.
