+++ devpack/src/main/java/io/neow3j/devpack/contracts/OracleContract.java
@@ -1,5 +1,6 @@
-package io.neow3j.devpack;
+package io.neow3j.devpack.contracts;
 
+import io.neow3j.devpack.ContractInterface;
 import io.neow3j.devpack.annotations.ContractHash;
 
 @ContractHash("0xfe924b7cfe89ddd271abaf7210a80a7e11178758")
